LayoutTests/fast/media/use-system-appearance-user-stylesheet-parsing-expected.txt

PASS getComputedStyle(test).color is "rgb(0, 0, 0)"
PASS getComputedStyle(test).color is "rgb(0, 128, 0)"
PASS getComputedStyle(test).color is "rgb(0, 0, 0)"
PASS successfullyParsed is true

TEST COMPLETE
